requirements.txt
streamlit
pandas
tensorflows